template/server/requirements.txt
# EmotionMate — Pipecat voice agent (Windows/Mac/Linux)
# Based on Oumi hackathon voice-agent template: https://github.com/oumi-ai/hack-with-oumi

pipecat-ai[whisper,kokoro]>=0.0.108
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
loguru>=0.7.0
openai>=1.0.0
httpx>=0.27.0
transformers>=4.40.0
torch>=2.0.0
